requirements.txt
streamlit
whisper
torch==2.1.2
requests
fpdf
pydub
python-docx
spacy==3.6.1
thinc==8.1.10
langchain
langchain-google-genai
openai-whisper
ffmpeg
numpy==1.24.4
en-core-sci-md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_md-0.5.3.tar.gz
